+++ NEWS.md
@@ -2,12 +2,12 @@
 
 * Bug fixed: the function `inversionSwappingTwoCircles` did not work for all cases.
 
-* New functions: `inversionFromCircle` and `midCircles`.
+* New functions: `inversionFromCircle`, `midCircles`, `CircleAB`.
 
-* New methods for `Circle` class: `tangentsThroughExternalPoint`, `isOrthogonal`
+* New methods for `Circle` class: `tangentsThroughExternalPoint`,`isOrthogonal`
 
 * New methods for `Triangle` class: `symmedialTriangle`, `symmedianPoint`, 
-`BrocardCircle`
+`BrocardCircle`, `LemoineCircleI/II/III`, `LemoineTriangle`, `ParryCircle`
 
 
 # PlaneGeometry 1.0.0 (2020-02-01)