+++ NEWS.md
@@ -4,13 +4,13 @@
 
 * New functions: `inversionFromCircle`, `midCircles`, `CircleAB`.
 
-* New methods for `Circle` class: `tangentsThroughExternalPoint`,`isOrthogonal`
+* New methods for `Circle` class: `tangentsThroughExternalPoint`,`isOrthogonal`.
 
 * New methods for `Triangle` class: `symmedialTriangle`, `symmedianPoint`, 
 `BrocardCircle`, `BrocardPoints`, `LemoineCircleI/II/III`, `LemoineTriangle`, 
-`ParryCircle`, `SteinerEllipse`
+`ParryCircle`, `SteinerEllipse`.
 
-* New method for `Line` class: `distance`
+* New methods for `Line` class: `distance`, `parallel`.
 
 
 # PlaneGeometry 1.0.0 (2020-02-01)